// src/EngagementOrganizer.SPA/src/app/api/EngagementOrganizerApiClient/model/utilization.ts
/**
 * Engagement Organizer
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { UtilizationRow } from './utilizationRow';

export interface Utilization { 
    utilizationMonths?: Array<UtilizationRow>;
    utilizationQuarter?: Array<UtilizationRow>;
    utilizationHalf?: Array<UtilizationRow>;
    utilizationYear?: UtilizationRow;
}